update booking system

+++ backend/config/cloudinary.js
@@ -13,11 +13,15 @@ const uploadOnCloudinary = async (filepath) => {
             return null
         }
         const uploadResult = await cloudinary.uploader.upload(filepath)
-        fs.unlinkSync(filepath)
+        if (fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath)
+        }
         return uploadResult.secure_url
         
     } catch (error) {
-        fs.unlinkSync(filepath)
+        if (filepath && fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath)
+        }
         console.log(error)
         
     }
